// src/app/models/enums/sort-type.ts
export enum SortTypeEnum {
  SORT_FOLDERS = 'sortFolders',
  SORT_SECTIONS = 'sortSections',
  SORT_SUBSECTIONS = 'sortSubsections',
  SORT_CARDS = 'sortCards',
  SORT_CARD_NOTES = 'sortCardNotes',
  SORT_NOTEBOOK_SECTIONS = 'sortNotebookSections',
  SORT_NOTEBOOK_CARDS = 'sortNotebookCards',
  SORT_TIMELINE_EVENT_NOTES = 'sortTimelineEventNotes',
  SORT_NARRATIVE_STORIES = 'sortNarrativeStories',
  SORT_TASKS = 'sortTasks',
}